backend/requirements.txt
beautifulsoup4==4.12.2
Chroma==0.2.0
fastapi==0.102.0
langchain==0.0.273
langsmith==0.0.26
nltk==3.8.1
numpy==1.25.2
openai==0.27.9
Pillow==10.0.0
playwright==1.37.0
pydantic==2.3.0
pypdf==3.15.2
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
soupsieve==2.4.1
unstructured==0.10.6
uvicorn==0.23.2
Development
loguru==0.7.0
pytest==7.4.0